refactor: CalendarView 리팩토링 전
