+++ DangProject/Presentation/Home/VIewModel/GraphVIewModel/GraphViewModel.swift
@@ -15,13 +15,12 @@ struct weekTemp {
     
     init(tempNutrient: tempNutrient) {
         guard let dangdang = tempNutrient.dang else { return }
-        // MARK: 여기서 형변환 ㄴㄴ
         guard let dangdang = Double(dangdang) else { return }
         self.dangdang = dangdang
     }
 }
 
-class GraphItemViewModel: ViewModelFactoryProtocol {
+class GraphViewModel: ViewModelFactoryProtocol {
     var items = BehaviorRelay<[weekTemp]>(value: [])
     
     init(item: [weekTemp]) {
