+++ DangProject/Presentation/Home/View/Battery/Calendar/CalendarCollectionViewCell.swift
@@ -1,22 +1,31 @@
 //
-//  CalendarCollectionViewCell.swift
+//  CalendarCollectionView.swift
 //  DangProject
 //
-//  Created by 김동우 on 2022/02/17.
+//  Created by 김동우 on 2022/02/23.
 //
 
-import Foundation
 import UIKit
+import RxSwift
 
 class CalendarCollectionViewCell: UICollectionViewCell {
     static let identifier = "CalendarCollectionViewCell"
-    
-    var dayLabel = UILabel()
+    private var viewModel: CalendarViewModel?
+    lazy var calendarCollectionView: UICollectionView = {
+        let layout = UICollectionViewFlowLayout()
+        layout.minimumInteritemSpacing = 0
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        
+        return collectionView
+    }()
+    private var testAnimationButton = UIButton()
+    private var disposeBag = DisposeBag()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
         configure()
-        layout()
+        autoLayout()
+        self.bringSubviewToFront(testAnimationButton)
     }
     
     required init?(coder: NSCoder) {
@@ -24,21 +33,67 @@ class CalendarCollectionViewCell: UICollectionViewCell {
     }
     
     private func configure() {
-        dayLabel.do {
-            $0.textColor = .black
-            $0.textAlignment = .center
-            $0.font = UIFont.boldSystemFont(ofSize: 15)
+        calendarCollectionView.do {
+            $0.register(CalendarCollectionViewCell.self, forCellWithReuseIdentifier: CalendarCollectionViewCell.identifier)
+            $0.dataSource = self
+            $0.delegate = self
+            $0.backgroundColor = .clear
         }
-        self.contentView.backgroundColor = .clear
     }
     
-    private func layout() {
-        [ dayLabel ].forEach() { self.addSubview($0) }
+    private func autoLayout() {
+        [ calendarCollectionView ].forEach() { self.addSubview($0) }
         
-        dayLabel.do {
+        calendarCollectionView.do {
             $0.translatesAutoresizingMaskIntoConstraints = false
             $0.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
-            $0.centerXAnchor.constraint(equalTo: self.centerXAnchor).isActive = true
+            $0.leadingAnchor.constraint(equalTo: self.leadingAnchor).isActive = true
+            $0.trailingAnchor.constraint(equalTo: self.trailingAnchor).isActive = true
+            $0.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
         }
     }
+    
+    func bind(viewModel: CalendarViewModel) {
+        self.viewModel = viewModel
+        subscribe()
+    }
+    
+    private func subscribe() {}
 }
+
+extension CalendarView: UICollectionViewDelegate, UICollectionViewDataSource {
+    func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return 1
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        numberOfItemsInSection section: Int) -> Int {
+        return viewModel?.calendarData.value.calendar?.days?.count ?? 0
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: 55, height: 50)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,
+                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CalendarCollectionViewCell.identifier, for: indexPath) as? CalendarCollectionViewCell else { return UICollectionViewCell() }
+        
+        if let data = viewModel?.calendarData.value.calendar?.days?[indexPath.item] {
+            let calendarCellViewModel = CalendarCellViewModel(calendarData: CalendarCellEntity(days: data))
+            cell.bind(viewModel: calendarCellViewModel)
+        }
+        return cell
+    }
+}
+
+extension CalendarView: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        insetForSectionAt section: Int) -> UIEdgeInsets {
+        return UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
+    }
+}
+
