--- DangProject/Presentation/Home/View/Battery/Calendar/CalendarView.swift
@@ -1,102 +1,0 @@
-//
-//  CalendarCollectionView.swift
-//  DangProject
-//
-//  Created by 김동우 on 2022/02/23.
-//
-
-import UIKit
-import RxSwift
-
-class CalendarView: UIView, ViewFactoryProtocol {
-    static let identifier = "CalendarCollectionView"
-    var viewModel: CalendarViewModel?
-    lazy var calendarCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.minimumInteritemSpacing = 0
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        
-        return collectionView
-    }()
-    var testAnimationButton = UIButton()
-    var disposeBag = DisposeBag()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-        autoLayout()
-        self.bringSubviewToFront(testAnimationButton)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configure() {
-        calendarCollectionView.do {
-            $0.register(CalendarCollectionViewCell.self, forCellWithReuseIdentifier: CalendarCollectionViewCell.identifier)
-            $0.dataSource = self
-            $0.delegate = self
-            $0.backgroundColor = .clear
-        }
-    }
-    
-    private func autoLayout() {
-        [ calendarCollectionView ].forEach() { self.addSubview($0) }
-        
-        calendarCollectionView.do {
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            $0.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
-            $0.leadingAnchor.constraint(equalTo: self.leadingAnchor).isActive = true
-            $0.trailingAnchor.constraint(equalTo: self.trailingAnchor).isActive = true
-            $0.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
-        }
-    }
-    
-    func bind(viewModel: CalendarViewModel) {
-        self.viewModel = viewModel
-        subscribe()
-    }
-    
-    private func subscribe() {
-        viewModel?.calendarData
-            .subscribe { data in
-                print("값이 변했구나~")
-            }
-            .disposed(by: disposeBag)
-    }
-}
-
-extension CalendarView: UICollectionViewDelegate, UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        numberOfItemsInSection section: Int) -> Int {
-        return viewModel?.calendarData.value.calendar?.days?.count ?? 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 55, height: 50)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CalendarCollectionViewCell.identifier, for: indexPath) as? CalendarCollectionViewCell else { return UICollectionViewCell() }
-        guard let day = viewModel?.calendarData.value.calendar?.days else { return UICollectionViewCell() }
-        cell.dayLabel.text = day[indexPath.item]
-        return cell
-    }
-}
-
-extension CalendarView: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-    }
-}
-
